Moves more services and registrations "up", removes AspNetCoreComposer logic and moves to AddWebComponents
