+++ src/Umbraco.Core/Cache/DistributedCache.cs
@@ -164,9 +164,15 @@ namespace Umbraco.Web.Cache
         #endregion
 
         // helper method to get an ICacheRefresher by its unique identifier
-        private  ICacheRefresher GetRefresherById(Guid refresherGuid)
+        private ICacheRefresher GetRefresherById(Guid refresherGuid)
         {
-            return _cacheRefreshers[refresherGuid];
+            ICacheRefresher refresher = _cacheRefreshers[refresherGuid];
+            if (refresher == null)
+            {
+                throw new InvalidOperationException($"No cache refresher found with id {refresherGuid}");
+            }
+
+            return refresher;
         }
     }
 }
