+++ src/Umbraco.Core/DependencyInjection/UmbracoBuilder.Collections.cs
@@ -1,11 +1,21 @@
+using System.Security.Cryptography;
+using Umbraco.Core.Cache;
 using Umbraco.Core.Composing;
+using Umbraco.Core.Dashboards;
 using Umbraco.Core.HealthCheck;
 using Umbraco.Core.Manifest;
+using Umbraco.Core.PackageActions;
 using Umbraco.Core.PropertyEditors;
+using Umbraco.Core.PropertyEditors.Validators;
+using Umbraco.Core.Strings;
+using Umbraco.Core.Trees;
 using Umbraco.Web.Actions;
 using Umbraco.Web.ContentApps;
 using Umbraco.Web.Dashboards;
 using Umbraco.Web.Editors;
+using Umbraco.Web.HealthCheck;
+using Umbraco.Web.HealthCheck.NotificationMethods;
+using Umbraco.Web.Media.EmbedProviders;
 using Umbraco.Web.Routing;
 using Umbraco.Web.Sections;
 using Umbraco.Web.Tour;
@@ -17,6 +27,97 @@ namespace Umbraco.Core.DependencyInjection
     /// </summary>
     public static partial class UmbracoBuilderExtensions
     {
+        /// <summary>
+        /// Adds all core collection builders
+        /// </summary>
+        internal static void AddAllCoreCollectionBuilders(this IUmbracoBuilder builder)
+        {
+            builder.CacheRefreshers().Add(() => builder.TypeLoader.GetCacheRefreshers());
+            builder.DataEditors().Add(() => builder.TypeLoader.GetDataEditors());
+            builder.Actions().Add(() => builder.TypeLoader.GetTypes<IAction>());
+            // register known content apps
+            builder.ContentApps()
+                .Append<ListViewContentAppFactory>()
+                .Append<ContentEditorContentAppFactory>()
+                .Append<ContentInfoContentAppFactory>()
+                .Append<ContentTypeDesignContentAppFactory>()
+                .Append<ContentTypeListViewContentAppFactory>()
+                .Append<ContentTypePermissionsContentAppFactory>()
+                .Append<ContentTypeTemplatesContentAppFactory>();
+            // all built-in finders in the correct order,
+            // devs can then modify this list on application startup
+            builder.ContentFinders()                
+                .Append<ContentFinderByPageIdQuery>()
+                .Append<ContentFinderByUrl>()
+                .Append<ContentFinderByIdPath>()
+                /*.Append<ContentFinderByUrlAndTemplate>() // disabled, this is an odd finder */
+                .Append<ContentFinderByUrlAlias>()
+                .Append<ContentFinderByRedirectUrl>();
+            builder.EditorValidators().Add(() => builder.TypeLoader.GetTypes<IEditorValidator>());
+            builder.HealthChecks().Add(() => builder.TypeLoader.GetTypes<Core.HealthCheck.HealthCheck>());
+            builder.HealthCheckNotificationMethods().Add(() => builder.TypeLoader.GetTypes<IHealthCheckNotificationMethod>());
+            builder.TourFilters();
+            builder.UrlProviders()
+                .Append<AliasUrlProvider>()
+                .Append<DefaultUrlProvider>();
+            builder.MediaUrlProviders()
+                .Append<DefaultMediaUrlProvider>();
+            // register back office sections in the order we want them rendered
+            builder.Sections()
+                .Append<ContentSection>()
+                .Append<MediaSection>()
+                .Append<SettingsSection>()
+                .Append<PackagesSection>()
+                .Append<UsersSection>()
+                .Append<MembersSection>()
+                .Append<FormsSection>()
+                .Append<TranslationSection>();
+            builder.Components();
+            // register core CMS dashboards and 3rd party types - will be ordered by weight attribute & merged with package.manifest dashboards
+            builder.Dashboards()
+                .Add<ContentDashboard>()
+                .Add<ExamineDashboard>()
+                .Add<FormsDashboard>()
+                .Add<HealthCheckDashboard>()
+                .Add<ManifestDashboard>()
+                .Add<MediaDashboard>()
+                .Add<MembersDashboard>()
+                .Add<ProfilerDashboard>()
+                .Add<PublishedStatusDashboard>()
+                .Add<RedirectUrlDashboard>()
+                .Add<SettingsDashboard>()
+                .Add(builder.TypeLoader.GetTypes<IDashboard>());
+            builder.PackageActions().Add(() => builder.TypeLoader.GetPackageActions());
+            builder.DataValueReferenceFactories();
+            builder.PropertyValueConverters().Append(builder.TypeLoader.GetTypes<IPropertyValueConverter>());
+            builder.UrlSegmentProviders().Append<DefaultUrlSegmentProvider>();
+            builder.ManifestValueValidators()
+                .Add<RequiredValidator>()
+                .Add<RegexValidator>()
+                .Add<DelimitedValueValidator>()
+                .Add<EmailValidator>()
+                .Add<IntegerValidator>()
+                .Add<DecimalValidator>();
+            builder.ManifestFilters();
+            builder.MediaUrlGenerators();
+            // register OEmbed providers - no type scanning - all explicit opt-in of adding types, IEmbedProvider is not IDiscoverable
+            builder.OEmbedProviders()
+                .Append<YouTube>()
+                .Append<Twitter>()
+                .Append<Vimeo>()
+                .Append<DailyMotion>()
+                .Append<Flickr>()
+                .Append<Slideshare>()
+                .Append<Kickstarter>()
+                .Append<GettyImages>()
+                .Append<Ted>()
+                .Append<Soundcloud>()
+                .Append<Issuu>()
+                .Append<Hulu>()
+                .Append<Giphy>();
+            builder.SearchableTrees().Add(() => builder.TypeLoader.GetTypes<ISearchableTree>());
+        }
+
         /// <summary>
         /// Gets the actions collection builder.
         /// </summary>
@@ -51,6 +152,9 @@ namespace Umbraco.Core.DependencyInjection
         /// <param name="builder">The builder.</param>
         public static HealthCheckCollectionBuilder HealthChecks(this IUmbracoBuilder builder)
             => builder.WithCollectionBuilder<HealthCheckCollectionBuilder>();
+
+        public static HealthCheckNotificationMethodCollectionBuilder HealthCheckNotificationMethods(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<HealthCheckNotificationMethodCollectionBuilder>();
 
         /// <summary>
         /// Gets the TourFilters collection builder.
@@ -90,18 +194,63 @@ namespace Umbraco.Core.DependencyInjection
         /// </summary>
         /// <param name="builder">The builder.</param>
         public static DashboardCollectionBuilder Dashboards(this IUmbracoBuilder builder)
-            => builder.WithCollectionBuilder<DashboardCollectionBuilder>()
-                .Add<ContentDashboard>()
-                .Add<ExamineDashboard>()
-                .Add<FormsDashboard>()
-                .Add<HealthCheckDashboard>()
-                .Add<ManifestDashboard>()
-                .Add<MediaDashboard>()
-                .Add<MembersDashboard>()
-                .Add<ProfilerDashboard>()
-                .Add<PublishedStatusDashboard>()
-                .Add<RedirectUrlDashboard>()
-                .Add<SettingsDashboard>();
+            => builder.WithCollectionBuilder<DashboardCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the cache refreshers collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        public static CacheRefresherCollectionBuilder CacheRefreshers(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<CacheRefresherCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the package actions collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        internal static PackageActionCollectionBuilder PackageActions(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<PackageActionCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the data editor collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        public static DataEditorCollectionBuilder DataEditors(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<DataEditorCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the data value reference factory collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        public static DataValueReferenceFactoryCollectionBuilder DataValueReferenceFactories(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<DataValueReferenceFactoryCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the property value converters collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        public static PropertyValueConverterCollectionBuilder PropertyValueConverters(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<PropertyValueConverterCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the url segment providers collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        public static UrlSegmentProviderCollectionBuilder UrlSegmentProviders(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<UrlSegmentProviderCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the validators collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        internal static ManifestValueValidatorCollectionBuilder ManifestValueValidators(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<ManifestValueValidatorCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the manifest filter collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        public static ManifestFilterCollectionBuilder ManifestFilters(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<ManifestFilterCollectionBuilder>();
 
         /// <summary>
         /// Gets the content finders collection builder.
@@ -109,5 +258,18 @@ namespace Umbraco.Core.DependencyInjection
         /// <param name="builder">The builder.</param>
         public static MediaUrlGeneratorCollectionBuilder MediaUrlGenerators(this IUmbracoBuilder builder)
             => builder.WithCollectionBuilder<MediaUrlGeneratorCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the backoffice OEmbed Providers collection builder.
+        /// </summary>
+        /// <param name="builder">The builder.</param>
+        public static EmbedProvidersCollectionBuilder OEmbedProviders(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<EmbedProvidersCollectionBuilder>();
+
+        /// <summary>
+        /// Gets the back office searchable tree collection builder
+        /// </summary>
+        public static SearchableTreeCollectionBuilder SearchableTrees(this IUmbracoBuilder builder)
+            => builder.WithCollectionBuilder<SearchableTreeCollectionBuilder>();
     }
 }
