+++ src/Umbraco.Core/Media/IImageDimensionExtractor.cs
