+++ src/Umbraco.Core/Media/ImageSize.cs
