+++ src/Umbraco.Core/Media/UploadAutoFillProperties.cs
