+++ src/Umbraco.Core/PropertyEditors/TextStringValueConverter.cs
