+++ src/Umbraco.Infrastructure/Search/UmbracoTreeSearcher.cs
@@ -8,11 +8,11 @@ using Umbraco.Core.Models;
 using Umbraco.Core.Models.Entities;
 using Umbraco.Core.Persistence;
 using Umbraco.Core.Services;
+using Umbraco.Core.Trees;
 using Umbraco.Examine;
 using Umbraco.Web.Models.ContentEditing;
 using Umbraco.Web.Models.Mapping;
 using Umbraco.Web.Routing;
-using Umbraco.Web.Trees;
 
 namespace Umbraco.Web.Search
 {
