+++ src/Umbraco.Web.BackOffice/Trees/ContentTreeController.cs
@@ -22,6 +22,7 @@ using Microsoft.Extensions.Options;
 using Umbraco.Web.Trees;
 using Microsoft.AspNetCore.Authorization;
 using Umbraco.Web.Common.Authorization;
+using Umbraco.Core.Trees;
 
 namespace Umbraco.Web.BackOffice.Trees
 {
