+++ src/Umbraco.Web.BackOffice/Trees/ContentTypeTreeController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Umbraco.Core;
 using Umbraco.Core.Models;
 using Umbraco.Core.Services;
+using Umbraco.Core.Trees;
 using Umbraco.Web.Actions;
 using Umbraco.Web.BackOffice.Filters;
 using Umbraco.Web.Common.Attributes;
