+++ src/Umbraco.Web.BackOffice/Trees/MediaTypeTreeController.cs
@@ -15,6 +15,7 @@ using Umbraco.Web.Trees;
 using Umbraco.Web.WebApi;
 using Microsoft.AspNetCore.Authorization;
 using Umbraco.Web.Common.Authorization;
+using Umbraco.Core.Trees;
 
 namespace Umbraco.Web.BackOffice.Trees
 {
