+++ src/Umbraco.Web.BackOffice/Trees/MemberTypeTreeController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Umbraco.Core;
 using Umbraco.Core.Models;
 using Umbraco.Core.Services;
+using Umbraco.Core.Trees;
 using Umbraco.Web.BackOffice.Filters;
 using Umbraco.Web.BackOffice.Trees;
 using Umbraco.Web.Common.Attributes;
