+++ src/Umbraco.Web.BackOffice/Trees/TreeControllerBase.cs
@@ -7,6 +7,7 @@ using Umbraco.Core.Events;
 using Umbraco.Core.Models;
 using Umbraco.Core.Models.Entities;
 using Umbraco.Core.Persistence;
+using Umbraco.Core.Trees;
 using Umbraco.Extensions;
 using Umbraco.Web.BackOffice.Controllers;
 using Umbraco.Web.Common.Filters;
