+++ src/Umbraco.Web.Common/DependencyInjection/UmbracoBuilderExtensions.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data.Common;
 using System.Data.SqlClient;
 using System.IO;
+using System.Linq;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using Microsoft.AspNetCore.Builder;
@@ -25,22 +26,39 @@ using Umbraco.Core.Configuration;
 using Umbraco.Core.Configuration.Models;
 using Umbraco.Core.Configuration.Models.Validation;
 using Umbraco.Core.DependencyInjection;
+using Umbraco.Core.Diagnostics;
+using Umbraco.Core.Hosting;
 using Umbraco.Core.IO;
 using Umbraco.Core.Logging;
 using Umbraco.Core.Persistence;
 using Umbraco.Core.Persistence.SqlSyntax;
 using Umbraco.Core.Runtime;
+using Umbraco.Core.Security;
 using Umbraco.Extensions;
 using Umbraco.Infrastructure.DependencyInjection;
 using Umbraco.Infrastructure.HostedServices;
 using Umbraco.Infrastructure.HostedServices.ServerRegistration;
+using Umbraco.Infrastructure.PublishedCache.DependencyInjection;
 using Umbraco.Infrastructure.Runtime;
+using Umbraco.Net;
 using Umbraco.Web.Cache;
 using Umbraco.Web.Common.ApplicationModels;
 using Umbraco.Web.Common.AspNetCore;
+using Umbraco.Web.Common.Controllers;
 using Umbraco.Web.Common.DependencyInjection;
+using Umbraco.Web.Common.Install;
+using Umbraco.Web.Common.Lifetime;
+using Umbraco.Web.Common.Macros;
+using Umbraco.Web.Common.Middleware;
+using Umbraco.Web.Common.ModelBinders;
 using Umbraco.Web.Common.Profiler;
+using Umbraco.Web.Common.Routing;
+using Umbraco.Web.Common.Security;
+using Umbraco.Web.Common.Templates;
+using Umbraco.Web.Macros;
+using Umbraco.Web.Security;
 using Umbraco.Web.Telemetry;
+using Umbraco.Web.Templates;
 using IHostingEnvironment = Umbraco.Core.Hosting.IHostingEnvironment;
 
 namespace Umbraco.Web.Common.DependencyInjection
@@ -123,58 +141,9 @@ namespace Umbraco.Web.Common.DependencyInjection
             ));
 
             builder.AddCoreInitialServices();
+
+            // TODO: This should be a separate call to opt-in to plugins
             builder.AddComposers();
-
-            return builder;
-        }
-
-        /// <summary>
-        /// Adds Umbraco composers for plugins
-        /// </summary>
-        public static IUmbracoBuilder AddComposers(this IUmbracoBuilder builder)
-        {
-            IEnumerable<Type> composerTypes = builder.TypeLoader.GetTypes<IComposer>();
-            IEnumerable<Attribute> enableDisable = builder.TypeLoader.GetAssemblyAttributes(typeof(EnableComposerAttribute), typeof(DisableComposerAttribute));
-            new Composers(builder, composerTypes, enableDisable, builder.BuilderLoggerFactory.CreateLogger<Composers>()).Compose();
-
-            return builder;
-        }
-
-        /// <summary>
-        /// Add Umbraco configuration services and options
-        /// </summary>
-        public static IUmbracoBuilder AddConfiguration(this IUmbracoBuilder builder)
-        {
-            // Register configuration validators.
-            builder.Services.AddSingleton<IValidateOptions<ContentSettings>, ContentSettingsValidator>();
-            builder.Services.AddSingleton<IValidateOptions<GlobalSettings>, GlobalSettingsValidator>();
-            builder.Services.AddSingleton<IValidateOptions<HealthChecksSettings>, HealthChecksSettingsValidator>();
-            builder.Services.AddSingleton<IValidateOptions<RequestHandlerSettings>, RequestHandlerSettingsValidator>();
-
-            // Register configuration sections.
-            builder.Services.Configure<ActiveDirectorySettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigActiveDirectory));
-            builder.Services.Configure<ConnectionStrings>(builder.Config.GetSection("ConnectionStrings"), o => o.BindNonPublicProperties = true);
-            builder.Services.Configure<ContentSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigContent));
-            builder.Services.Configure<CoreDebugSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigCoreDebug));
-            builder.Services.Configure<ExceptionFilterSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigExceptionFilter));
-            builder.Services.Configure<GlobalSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigGlobal));
-            builder.Services.Configure<HealthChecksSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigHealthChecks));
-            builder.Services.Configure<HostingSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigHosting));
-            builder.Services.Configure<ImagingSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigImaging));
-            builder.Services.Configure<IndexCreatorSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigExamine));
-            builder.Services.Configure<KeepAliveSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigKeepAlive));
-            builder.Services.Configure<LoggingSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigLogging));
-            builder.Services.Configure<MemberPasswordConfigurationSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigMemberPassword));
-            builder.Services.Configure<ModelsBuilderSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigModelsBuilder), o => o.BindNonPublicProperties = true);
-            builder.Services.Configure<NuCacheSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigNuCache));
-            builder.Services.Configure<RequestHandlerSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigRequestHandler));
-            builder.Services.Configure<RuntimeSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigRuntime));
-            builder.Services.Configure<SecuritySettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigSecurity));
-            builder.Services.Configure<TourSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigTours));
-            builder.Services.Configure<TypeFinderSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigTypeFinder));
-            builder.Services.Configure<UserPasswordConfigurationSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigUserPassword));
-            builder.Services.Configure<WebRoutingSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigWebRouting));
-            builder.Services.Configure<UmbracoPluginSettings>(builder.Config.GetSection(Core.Constants.Configuration.ConfigPlugins));
 
             return builder;
         }
@@ -195,8 +164,7 @@ namespace Umbraco.Web.Common.DependencyInjection
             return builder;
         }
 
-        // TODO: Not sure this needs to exist and/or be public?
-        public static IUmbracoBuilder AddHttpClients(this IUmbracoBuilder builder)
+        private static IUmbracoBuilder AddHttpClients(this IUmbracoBuilder builder)
         {
             builder.Services.AddHttpClient();
             return builder;
@@ -241,6 +209,9 @@ namespace Umbraco.Web.Common.DependencyInjection
             return builder;
         }
 
+        /// <summary>
+        /// Adds all web based services required for Umbraco to run
+        /// </summary>
         public static IUmbracoBuilder AddWebComponents(this IUmbracoBuilder builder)
         {
             // Add service session
@@ -256,6 +227,59 @@ namespace Umbraco.Web.Common.DependencyInjection
             builder.Services.TryAddEnumerable(ServiceDescriptor.Transient<IApplicationModelProvider, UmbracoApiBehaviorApplicationModelProvider>());
             builder.Services.TryAddEnumerable(ServiceDescriptor.Transient<IApplicationModelProvider, BackOfficeApplicationModelProvider>());
             builder.Services.AddUmbracoImageSharp(builder.Config);
+
+            // AspNetCore specific services
+            builder.Services.AddUnique<IHttpContextAccessor, HttpContextAccessor>();
+            builder.Services.AddUnique<IRequestAccessor, AspNetCoreRequestAccessor>();
+
+            // Our own netcore implementations
+            builder.Services.AddMultipleUnique<IUmbracoApplicationLifetimeManager, IUmbracoApplicationLifetime, AspNetCoreUmbracoApplicationLifetime>();
+
+            builder.Services.AddUnique<IApplicationShutdownRegistry, AspNetCoreApplicationShutdownRegistry>();
+
+            // The umbraco request lifetime
+            builder.Services.AddMultipleUnique<IUmbracoRequestLifetime, IUmbracoRequestLifetimeManager, UmbracoRequestLifetime>();
+
+            // Password hasher
+            builder.Services.AddUnique<IPasswordHasher, AspNetCorePasswordHasher>();
+
+            builder.Services.AddUnique<ICookieManager, AspNetCoreCookieManager>();
+            builder.Services.AddTransient<IIpResolver, AspNetCoreIpResolver>();
+            builder.Services.AddUnique<IUserAgentProvider, AspNetCoreUserAgentProvider>();
+
+            builder.Services.AddMultipleUnique<ISessionIdResolver, ISessionManager, AspNetCoreSessionManager>();
+
+            builder.Services.AddUnique<IMarchal, AspNetCoreMarchal>();
+
+            builder.Services.AddUnique<IProfilerHtml, WebProfilerHtml>();
+
+            builder.Services.AddUnique<IMacroRenderer, MacroRenderer>();
+            builder.Services.AddUnique<IMemberUserKeyProvider, MemberUserKeyProvider>();
+
+            // register the umbraco context factory
+
+            builder.Services.AddUnique<IUmbracoContextFactory, UmbracoContextFactory>();
+            builder.Services.AddUnique<IBackOfficeSecurityFactory, BackOfficeSecurityFactory>();
+            builder.Services.AddUnique<IBackOfficeSecurityAccessor, HybridBackofficeSecurityAccessor>();
+            builder.Services.AddUnique<IUmbracoWebsiteSecurityAccessor, HybridUmbracoWebsiteSecurityAccessor>();
+
+            var umbracoApiControllerTypes = builder.TypeLoader.GetUmbracoApiControllers().ToList();
+            builder.WithCollectionBuilder<UmbracoApiControllerTypeCollectionBuilder>()
+                .Add(umbracoApiControllerTypes);
+
+            builder.Services.AddUnique<InstallAreaRoutes>();
+
+            builder.Services.AddUnique<UmbracoRequestLoggingMiddleware>();
+            builder.Services.AddUnique<UmbracoRequestMiddleware>();
+            builder.Services.AddUnique<BootFailedMiddleware>();
+
+            builder.Services.AddUnique<UmbracoJsonModelBinder>();
+
+            builder.Services.AddUnique<ITemplateRenderer, TemplateRenderer>();
+            builder.Services.AddUnique<IPublicAccessChecker, PublicAccessChecker>();
+
+            builder.AddNuCache();
+            builder.AddHttpClients();
 
             return builder;
         }
