+++ src/Umbraco.Web.Common/Runtime/AspNetCoreComposer.cs
@@ -35,63 +35,5 @@ namespace Umbraco.Web.Common.Runtime
     [ComposeBefore(typeof(ICoreComposer))]
     public class AspNetCoreComposer : ComponentComposer<AspNetCoreComponent>, IComposer
     {
-        public override void Compose(IUmbracoBuilder builder)
-        {
-            base.Compose(builder);
-
-            // AspNetCore specific services
-            builder.Services.AddUnique<IHttpContextAccessor, HttpContextAccessor>();
-            builder.Services.AddUnique<IRequestAccessor, AspNetCoreRequestAccessor>();
-
-            // Our own netcore implementations
-            builder.Services.AddMultipleUnique<IUmbracoApplicationLifetimeManager, IUmbracoApplicationLifetime, AspNetCoreUmbracoApplicationLifetime>();
-
-            builder.Services.AddUnique<IApplicationShutdownRegistry, AspNetCoreApplicationShutdownRegistry>();
-
-            // The umbraco request lifetime
-            builder.Services.AddMultipleUnique<IUmbracoRequestLifetime, IUmbracoRequestLifetimeManager, UmbracoRequestLifetime>();
-
-            // Password hasher
-            builder.Services.AddUnique<IPasswordHasher, AspNetCorePasswordHasher>();
-
-            builder.Services.AddUnique<ICookieManager, AspNetCoreCookieManager>();
-            builder.Services.AddTransient<IIpResolver, AspNetCoreIpResolver>();
-            builder.Services.AddUnique<IUserAgentProvider, AspNetCoreUserAgentProvider>();
-
-            builder.Services.AddMultipleUnique<ISessionIdResolver, ISessionManager, AspNetCoreSessionManager>();
-
-            builder.Services.AddUnique<IMarchal, AspNetCoreMarchal>();
-
-            builder.Services.AddUnique<IProfilerHtml, WebProfilerHtml>();
-
-            builder.Services.AddUnique<IMacroRenderer, MacroRenderer>();
-            builder.Services.AddUnique<IMemberUserKeyProvider, MemberUserKeyProvider>();
-
-            // register the umbraco context factory
-
-            builder.Services.AddUnique<IUmbracoContextFactory, UmbracoContextFactory>();
-            builder.Services.AddUnique<IBackOfficeSecurityFactory, BackOfficeSecurityFactory>();
-            builder.Services.AddUnique<IBackOfficeSecurityAccessor, HybridBackofficeSecurityAccessor>();
-            builder.Services.AddUnique<IUmbracoWebsiteSecurityAccessor, HybridUmbracoWebsiteSecurityAccessor>();
-
-            //register the install components
-            builder.AddInstaller();
-
-            var umbracoApiControllerTypes = builder.TypeLoader.GetUmbracoApiControllers().ToList();
-            builder.WithCollectionBuilder<UmbracoApiControllerTypeCollectionBuilder>()
-                .Add(umbracoApiControllerTypes);
-
-
-            builder.Services.AddUnique<InstallAreaRoutes>();
-
-            builder.Services.AddUnique<UmbracoRequestLoggingMiddleware>();
-            builder.Services.AddUnique<UmbracoRequestMiddleware>();
-            builder.Services.AddUnique<BootFailedMiddleware>();
-
-            builder.Services.AddUnique<UmbracoJsonModelBinder>();
-
-            builder.Services.AddUnique<ITemplateRenderer, TemplateRenderer>();
-            builder.Services.AddUnique<IPublicAccessChecker, PublicAccessChecker>();            
-        }
     }
 }
